requirements.txt
python-gnupg==0.4.6
PyYAML==5.4.1
requests==2.25.1
schedule==0.6.0
tinycss2==1.1.0
